% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{average_difference}
\alias{average_difference}
\title{Compute Difference in "Healthy" and "Less Healthy" food options}
\usage{
average_difference(clean_data)
}
\arguments{
\item{clean_data}{The data retrieved from a cleaned version of the Qualtrics Survey.}
}
\value{
Average_Difference for the average price difference.
}
\description{
This function finds the average difference in price between the healthier and
less healthy food options for a store
}
\details{
This function is used to determine adequate replacements for the market basket prices.
}
\examples{
clean_data <- read_nemss(file)
find_difference(clean_data)

}
