% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{baked_goods_avail}
\alias{baked_goods_avail}
\title{Compute the availability points associated with low-fat baked goods}
\usage{
baked_goods_avail(lowfat_baked_goods)
}
\arguments{
\item{lowfat_baked_goods}{The number of low-fat baked goods.}
}
\value{
The NEMS-S points associated with availability of low-fat baked goods.
}
\description{
This function takes in the number of varieties of low-fat baked goods and returns the NEMS-S points associated with availability.
}
\details{
This function implements the scoring method described in Table X of the NEMS-S manual. "Low-fat baked goods" are baked goods with less than or equal to 3 grams of fat per serving.
}
\examples{
lowfat_baked_goods <- sample(0:3, 10)
baked_goods_avail(lowfat_baked_goods)
}
