% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{baked_goods_cost}
\alias{baked_goods_cost}
\title{compute baked goods cost points}
\usage{
baked_goods_cost(healthier_baked_goods_price, regular_baked_goods_price)
}
\arguments{
\item{healthier_baked_goods_price}{the price of lower fat baked goods}

\item{regular_baked_goods_price}{the price of normal fat baked goods}
}
\value{
the NEMS-S points associated with baked goods price
}
\description{
compute baked goods cost points
}
\examples{
healthier_baked_goods_price <- rnorm(10,3.2,.5)
regular_baked_goods_price <- rnorm(10,3.0,.3)
}
