% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bread_avail}
\alias{bread_avail}
\title{compute whole grain bread availability points}
\usage{
bread_avail(varieties_of_whole_grain_bread)
}
\arguments{
\item{varieties_of_whole_grain_bread}{number of types of whole grain bread available}
}
\value{
the NEMS-S points associated with whole grain bread availability
}
\description{
compute whole grain bread availability points
}
\examples{
varieties_of_whole_grain_bread <- sample(1:10, 10)
bread_avail(varieties_of_whole_grain_bread)
}
