% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bread_cost}
\alias{bread_cost}
\title{compute bread cost points}
\usage{
bread_cost(whole_wheat_bread_price, white_bread_price)
}
\arguments{
\item{whole_wheat_bread_price}{the cost of whole wheat bread}

\item{white_bread_price}{the cost of white bread}
}
\value{
the NEMS-S points associated with bread cost
}
\description{
compute bread cost points
}
\examples{
whole_wheat_bread_price <- rnorm(10,1.5,.5)
white_bread_price <- rnorm(10,1.5,.3)
}
