% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{calculate_nems_score}
\alias{calculate_nems_score}
\title{Compute Availability and Cost Scores for all variables}
\usage{
calculate_nems_score(nems_data, detail = FALSE)
}
\arguments{
\item{nems_data}{A dataframe or tibble created by reading the raw qualtrics data
with `read_nems`}
}
\value{
Scores for total Availability and Cost, as well as given the option
  if Detail = TRUE to see the score for each individual variable.
}
\description{
This function uses all functions previously created and scores assigned to each variable and adds them into one total score for both availability and cost.
}
\details{
This function implements the scoring method described in Measure 9
  of the NEMS-S Protocol.
}
\examples{
calculate_nems_score(nems_sample)

}
