% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cereal_avail}
\alias{cereal_avail}
\title{Compute healthier cereal availability points}
\usage{
cereal_avail(healthy_cereal_varieties)
}
\arguments{
\item{healthy_cereal_varieties}{The number of varieties of cereal with less than 7g of sugar per serving.}
}
\value{
The NEMS-S points associated with healthier cereal availability.
}
\description{
This function takes in the number of varieties of healthier cereal options and returns the NEMS-S points associated with availability.
}
\examples{
healthy_cereal_varieties <- sample(1:10, 10)
cereal_avail(healthy_cereal_varieties)
}
