% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cereal_cost}
\alias{cereal_cost}
\title{compute cereal cost points}
\usage{
cereal_cost(healthier_cereal_price, regular_cereal_price)
}
\arguments{
\item{healthier_cereal_price}{the price of healthier cereal (<7g of sugar per serving)}

\item{regular_cereal_price}{the price of regular cereal (>7g of sugar per serving)}
}
\value{
the NEMS-S points associated with cereal cost
}
\description{
compute cereal cost points
}
\examples{
healthier_cereal_price <- rnorm(10,4.1,.5)
regular_cereal_price <- rnorm(10,3.5,.3)
}
