% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{chips_avail}
\alias{chips_avail}
\title{compute baked chips availability points}
\usage{
chips_avail(baked_chips_varieties)
}
\arguments{
\item{baked_chips_varieties}{varieties of baked chips offered}
}
\value{
the NEMS-S points associated with baked chips availability
}
\description{
compute baked chips availability points
}
\examples{
baked_chips_varieties <- sample(1:10, 10)
chips_avail(baked_chips_varieties)
}
