% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{chips_cost}
\alias{chips_cost}
\title{Compute chips cost points}
\usage{
chips_cost(baked_chips_price, regular_chips_price)
}
\arguments{
\item{baked_chips_price}{The price of baked chips.}

\item{regular_chips_price}{The price of regular chips.}
}
\value{
The NEMS-S points associated with chips cost.
}
\description{
This function takes in the cost of baked and regular chips and compares them to return a NEMS-S score for the cost.
}
\examples{
baked_chips_price <- rnorm(10,2.5,.5)
regular_chips_price <- rnorm(10,2.8,.3)
}
