% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{drinks_cost}
\alias{drinks_cost}
\title{compute drinks cost points}
\usage{
drinks_cost()
}
\arguments{
\item{healthier_drinks_price}{the price of 100% juice drinks}

\item{regular_drinks_price}{the price of}
}
\description{
compute drinks cost points
}
