% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{frozen_dinners_avail}
\alias{frozen_dinners_avail}
\title{compute the availability points associated with frozen dinners}
\usage{
frozen_dinners_avail(frozen_dinner_varieties)
}
\arguments{
\item{frozen_dinner_varieties}{the number of frozen dinner varieties offered (only )}
}
\value{
the NEMS-S points associated with availability of frozen dinners
}
\description{
compute the availability points associated with frozen dinners
}
\examples{
frozen_dinner_varieties <- sample(0:3, 10)
frozen_dinners_avail(frozen_dinner_varieties)
}
