% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{frozen_dinners_cost}
\alias{frozen_dinners_cost}
\title{compute frozen dinners cost points}
\usage{
frozen_dinners_cost(
  healthier_frozen_dinners_price,
  regular_frozen_dinners_price
)
}
\arguments{
\item{healthier_frozen_dinners_price}{the price of healthier option for frozen dinners}

\item{regular_frozen_dinners_price}{the price of regular fat frozen dinners}
}
\value{
the NEMS-S points associated with frozen dinners price
}
\description{
compute frozen dinners cost points
}
\examples{
healthier_frozen_dinners_price <- rnorm(10,4.5,.5)
regular_frozen_dinners_price <- rnorm(10,5.1,.3)
}
