% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fruit_avail}
\alias{fruit_avail}
\title{compute fruit availability points}
\usage{
fruit_avail(varieties_of_fruit)
}
\value{
the NEMS-S points associated with fruit availability
}
\description{
compute fruit availability points
}
\examples{
varieties_of_fruit <- sample(1:10, 10)
fruit_avail(varieties_of_fruit)
}
