% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ground_beef_avail}
\alias{ground_beef_avail}
\title{compute the availability points associated with ground beef}
\usage{
ground_beef_avail(lean_beef_varieties)
}
\arguments{
\item{lean_beef_varieties}{the number of lean beef varieties (<10% fat)}
}
\value{
the NEMS-S points associated with ground beef availability
}
\description{
compute the availability points associated with ground beef
}
\examples{
lean_beef_varieties <- sample(1:4, 10)
ground_beef_avail(lean_beef_varieties)
}
