% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ground_beef_cost}
\alias{ground_beef_cost}
\title{compute ground beef cost points}
\usage{
ground_beef_cost(lean_beef_price, regular_beef_price)
}
\arguments{
\item{lean_beef_price}{the price of ground beef}

\item{regular_beef_price}{the price of normal fat ground beef}
}
\value{
the NEMS-S points associated with beef price
}
\description{
compute ground beef cost points
}
\examples{
lean_beef_price <- rnorm(10,5.8,.5)
regular_beef_price <- rnorm(10,5.5,.3)
ground_beef_cost(lean_beef_price, regular_beef_price)
}
