% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{juice_drinks_avail}
\alias{juice_drinks_avail}
\title{Compute the availability points associated with healthier juice drinks}
\usage{
juice_drinks_avail(healthy_juice_varieties)
}
\arguments{
\item{healthy_juice_varieties}{The number of 100% juice drinks available.}
}
\value{
The NEMS-S points associated with availability of healthy juice drinks.
}
\description{
This function takes in the number of varieties of 100% juice drinks and returns the NEMS-S points associated with availability.
}
\examples{
healthy_juice_varieties <- sample(0:3, 10)
juice_drinks_avail(healthy_juice_varieties)
}
