% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{juice_drinks_cost}
\alias{juice_drinks_cost}
\title{Compute juice drinks cost points}
\usage{
juice_drinks_cost(healthier_juice_drinks_price, regular_juice_drinks_price)
}
\arguments{
\item{healthier_juice_drinks_price}{The cost of 100% juice drinks.}

\item{regular_juice_drinks_price}{The cost of non 100% juice drinks.}
}
\value{
The NEMS_S points associated with juice drinks cost.
}
\description{
This function takes in the cost of healthier and regular juice and compares them to return a NEMS-S score for the cost.
}
\examples{
healthier_juice_drinks_price <- rnorm(10,4.1,.5)
regular_juice_drinks_price <- rnorm(10,4.1,.3)
}
