% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{milk_avail}
\alias{milk_avail}
\title{Compute milk availability points}
\usage{
milk_avail(lowfat_milk_varieties, whole_milk_varieties)
}
\arguments{
\item{lowfat_milk_varieties}{The number of low fat milk cartons available.}

\item{whole_milk_varieties}{The number of whole milk cartons available.}
}
\value{
The NEMS-S points associated with milk availability.
}
\description{
This function takes in the number of lowfat and whole milk varieties at a given store and returns the NEMS-S points associated with the availability.
}
\details{
This function implements the scoring method described in Measure 1 of the NEMS-S Protocol. "Low-fat milk" is considered skim milk or 1 percent fat, whichever is available.
}
\examples{
lowfat_milk_varieties <- rnorm(10,,)
whole_milk_varieties <- rnorm(10,,)
milk_avail(lowfat_milk_varieties, whole_milk_varieties)
}
