% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{milk_cost}
\alias{milk_cost}
\title{Compute milk cost points}
\usage{
milk_cost(lowfat_milk_price, whole_milk_price)
}
\arguments{
\item{lowfat_milk_price}{The price of low fat milk.}

\item{whole_milk_price}{The price of whole milk.}
}
\value{
The NEMS-S points associated with milk price.
}
\description{
This function takes in the cost of lowfat and whole milk and compares them to return a NEMS-S score for the cost.
}
\examples{
lowfat_milk_price <- rnorm(10,2.8,.5)
whole_milk_price <- rnorm(10,3.1,.3)
milk_cost(lowfat_milk_price, whole_milk_price)
}
