% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemsr.r
\docType{package}
\name{nemsr}
\alias{nemsr}
\title{nemsr: A package for working with NEMS-S survey data}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

\section{Documentation and References}{

All the information for the scoring system and classification of "healthier" options come from the NEMS-S Protocol Measures and the Lundsford 2016 Thesis (most information comes from Table 4.1).
}

