% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{read_nemss}
\alias{read_nemss}
\title{Import and Clean Qualtrics NEMS_S Dataset}
\usage{
read_nemss(file)
}
\arguments{
\item{file}{The data file retrieved from the Qualtrics survey.}
}
\value{
A cleaned R dataframe with availability values and cost values for each variable.
}
\description{
This function takes the output of the Qualtrics survey and returns it as a
cleaned R dataframe.
}
\examples{
file <- system.file("extdata", "example_nems.sav", package = "nemsr")
read_nemss(file)

}
