% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{vegetable_avail}
\alias{vegetable_avail}
\title{compute vegetable availability points}
\usage{
vegetable_avail(varieties_of_vegetables)
}
\arguments{
\item{varieties_of_vegetables}{number of types of vegetables available}
}
\value{
the NEMS-S points associated with vegetables availability
}
\description{
compute vegetable availability points
}
\examples{
varieties_of_vegetables <- sample(1:10, 10)
vegetable_avail(varieties_of_vegetables)
}
