% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{wieners_avail}
\alias{wieners_avail}
\title{Compute the availability points associated with wieners}
\usage{
wieners_avail(fat_free_wieners, light_wieners)
}
\arguments{
\item{fat_free_wieners}{(fat content?)}

\item{light_wieners}{(specific fat content?)}
}
\value{
the NEMS-S points associated with availability of fat free and light fat wieners
}
\description{
This function takes in the .... this needs to be finished
}
\examples{
fat_free_wieners <- sample()
light_wieners <- sample()
wieners_avail(fat_free_wieners, light_wieners)
}
