% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{wieners_cost}
\alias{wieners_cost}
\title{compute wieners cost points}
\usage{
wieners_cost(lean_wieners_price, regular_wieners_price)
}
\arguments{
\item{lean_wieners_price}{the price of hot dogs with less fat (specific fat content?)}

\item{regular_wieners_price}{the price of normal hot dogs}
}
\value{
the NEMS-S points associated with wiener price
}
\description{
compute wieners cost points
}
\examples{
lean_wieners_price <- rnorm(10,2.5,.3)
regular_wieners_price <- rnorm(10,1.4,.5)
wieners_cost(lean_wieners_price,regular_wieners_price)
}
