% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{baked_goods_avail}
\alias{baked_goods_avail}
\title{compute the availability points associated with low-fat baked goods}
\usage{
baked_goods_avail(lowfat_baked_goods)
}
\arguments{
\item{lowfat_baked_goods}{number of low-fat baked goods}
}
\value{
the NEMS-S points associated with availability of low-fat baked goods
}
\description{
compute the availability points associated with low-fat baked goods
}
\examples{
lowfat_baked_goods <- sample(0:3, 10)
baked_goods_avail(lowfat_baked_goods)
}
