% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{baked_goods_cost}
\alias{baked_goods_cost}
\title{Compute baked goods cost points}
\usage{
baked_goods_cost(healthier_baked_goods_price, regular_baked_goods_price)
}
\arguments{
\item{healthier_baked_goods_price}{The price of lower fat baked goods (fat content).}

\item{regular_baked_goods_price}{The price of normal fat baked goods.}
}
\value{
The NEMS-S points associated with baked goods price.
}
\description{
This function takes in the cost of healthier and regular baked goods and compares them to return a NEMS-S score for the cost.
}
\examples{
healthier_baked_goods_price <- rnorm(10,3.2,.5)
regular_baked_goods_price <- rnorm(10,3.0,.3)
}
