% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{bread_avail}
\alias{bread_avail}
\title{Compute whole grain bread availability points}
\usage{
bread_avail(varieties_of_whole_grain_bread)
}
\arguments{
\item{varieties_of_whole_grain_bread}{The number of types of whole grain bread available.}
}
\value{
The NEMS-S points associated with whole grain bread availability.
}
\description{
This function takes in the number of varieties of whole grain bread and returns the NEMS-S points associated with availability.
}
\details{
This function implements the scoring method described in Measure 9 of the NEMS-S Protocol. "Whole grain bread" is 100% whole wheat and whole grain bread, and it is the healthier option.
}
\examples{
varieties_of_whole_grain_bread <- sample(1:10, 10)
bread_avail(varieties_of_whole_grain_bread)
}
