% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_scores.R
\name{bread_cost}
\alias{bread_cost}
\title{Compute bread cost points}
\usage{
bread_cost(whole_wheat_bread_price, white_bread_price)
}
\arguments{
\item{whole_wheat_bread_price}{The cost of whole wheat bread.}

\item{white_bread_price}{The cost of white bread.}
}
\value{
The NEMS-S points associated with bread cost.
}
\description{
This function takes in the cost of whole wheat bread and white bread and compares them to return a NEMS-S score for the cost.
}
\details{
This function implements the scoring method described in Table 4.1 of Lundsford (2016). "Whole grain bread" is 100% whole wheat and whole grain bread, and it is the healthier option. "White bread" is the regular option, bread made with refined flour.
}
\examples{
whole_wheat_bread_price <- rnorm(10,1.5,.5)
white_bread_price <- rnorm(10,1.5,.3)
}
