% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{bread_cost}
\alias{bread_cost}
\title{Compute bread cost points}
\usage{
bread_cost(whole_wheat_bread_price, white_bread_price)
}
\arguments{
\item{whole_wheat_bread_price}{The cost of whole wheat bread.}

\item{white_bread_price}{The cost of white bread.}
}
\value{
The NEMS-S points associated with bread cost.
}
\description{
This function takes in the cost of whole wheat bread and white bread and compares them to return a NEMS-S score for the cost.
}
\examples{
whole_wheat_bread_price <- rnorm(10,1.5,.5)
white_bread_price <- rnorm(10,1.5,.3)
}
