% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{calculate_market_basket}
\alias{calculate_market_basket}
\title{Compute market basket cost at a store}
\usage{
calculate_market_basket(nems_data)
}
\arguments{
\item{nems_data}{A dataframe or tibble containing one row per store, cleaned
from raw qualtrics survey data with the read_nemss function}
}
\value{
A tibble containing the weighted price for several types of goods
  including dairy, grain, etc. An additional column is appended with the
  number of replacements with mean values that had to be made in each type.
  A Total column  contains the cost of filling the market basket in that store.
}
\description{
Compute market basket cost at a store
}
\details{
This function applies the market basket outline found in the
  [Thrifty Food Plan, 2021](https://fns-prod.azureedge.us/sites/default/files/resource-files/TFP2021.pdf)
  Some adjustments needed to be made because the NEMS-S survey is mostly
  focused on the availability of low-calorie and low-fat options, and the
  market basket is based more on the price of observed consumer purchases. For
  example, the market basket includes a certain amount of poultry, but because
  low-fat poultry is not really a thing in the way that lean beef is, the NEMS-S
  instrument does not collect the price of poultry.

  Additionally, if a price was not available, we replaced it with the average
  price for all stores in the dataset.
}
\examples{
calculate_market_basket(nems_sample)

}
