% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cereal_avail}
\alias{cereal_avail}
\title{compute healthier cereal availability points}
\usage{
cereal_avail(healthy_cereal_varieties)
}
\arguments{
\item{healthy_cereal_varieties}{number of varieties of cereal with less than 7g of sugar per serving}
}
\value{
the NEMS-S points associated with healthier cereal availability
}
\description{
compute healthier cereal availability points
}
\examples{
healthy_cereal_varieties <- sample(1:10, 10)
cereal_avail(healthy_cereal_varieties)
}
