% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{cereal_avail}
\alias{cereal_avail}
\title{Compute healthier cereal availability points}
\usage{
cereal_avail(healthier_cereal_varieties)
}
\arguments{
\item{healthier_cereal_varieties}{The number of varieties of cereal with less than 7g of sugar per serving.}
}
\value{
The NEMS-S points associated with healthier cereal availability.
}
\description{
This function takes in the number of varieties of healthier cereal options and returns the NEMS-S points associated with availability.
}
\details{
This function implements the scoring method described in Measure 11 of the NEMS-S Protocol. "Healthier cereal" has less than 7g of sugar per serving.
}
\examples{
healthier_cereal_varieties <- sample(1:10, 10)
cereal_avail(healthier_cereal_varieties)
}
