% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{chips_avail}
\alias{chips_avail}
\title{Compute baked chips availability points}
\usage{
chips_avail(baked_chips_varieties)
}
\arguments{
\item{baked_chips_varieties}{The number of varieties of baked chips offered.}
}
\value{
The NEMS-S points associated with baked chips availability.
}
\description{
This function takes in the number of varieties of baked chips and returns the NEMS-S points associated with availability.
}
\examples{
baked_chips_varieties <- sample(1:10, 10)
chips_avail(baked_chips_varieties)
}
