% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_scores.R
\name{chips_avail}
\alias{chips_avail}
\title{Compute lowfat chips availability points}
\usage{
chips_avail(lowfat_chips_varieties)
}
\arguments{
\item{lowfat_chips_varieties}{The number of varieties of lowfat chips offered.}
}
\value{
The NEMS-S points associated with lowfat chips availability.
}
\description{
This function takes in the number of varieties of lowfat chips and returns the NEMS-S points associated with availability.
}
\details{
This function implements the scoring method described in Measure 10 of the NEMS-S Protocol. "Lowfat chips" are chips with less than or equal to 3g of fat per one ounce serving.
}
\examples{
lowfat_chips_varieties <- sample(1:10, 10)
chips_avail(lowfat_chips_varieties)
}
