% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{chips_cost}
\alias{chips_cost}
\title{compute chips cost points}
\usage{
chips_cost(baked_chips_price, regular_chips_price)
}
\arguments{
\item{baked_chips_price}{the price of baked chips}

\item{regular_chips_price}{the price of regular chips}
}
\value{
the NEMS-S points associated with chips
}
\description{
compute chips cost points
}
\examples{
baked_chips_price <- rnorm(10,2.5,.5)
regular_chips_price <- rnorm(10,2.8,.3)
}
