% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_scores.R
\name{chips_cost}
\alias{chips_cost}
\title{Compute chips cost points}
\usage{
chips_cost(lowfat_chips_price, regular_chips_price)
}
\arguments{
\item{lowfat_chips_price}{The price of lowfat chips.}

\item{regular_chips_price}{The price of regular chips.}
}
\value{
The NEMS-S points associated with chips cost.
}
\description{
This function takes in the cost of lowfat and regular chips and compares them to return a NEMS-S score for the cost.
}
\details{
This function implements the scoring method described in Table 4.1 of Lundsford (2016). "Lowfat chips" contain less than or equal to 3g of fat per 1oz serving.
}
\examples{
lowfat_chips_price <- rnorm(10,2.5,.5)
regular_chips_price <- rnorm(10,2.8,.3)
}
