% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{frozen_dinners_avail}
\alias{frozen_dinners_avail}
\title{Compute the availability points associated with frozen dinners}
\usage{
frozen_dinners_avail(frozen_dinner_varieties)
}
\arguments{
\item{frozen_dinner_varieties}{The number of frozen dinner varieties offered.}
}
\value{
The NEMS-S points associated with availability of frozen dinners.
}
\description{
This function takes in the number of varieties of frozen dinners and returns the NEMS-S points associated with availability.
}
\details{
This function implements the scoring method described in Table X of the NEMS-S manual.
}
\examples{
frozen_dinner_varieties <- sample(0:3, 10)
frozen_dinners_avail(frozen_dinner_varieties)
}
