% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{frozen_dinners_cost}
\alias{frozen_dinners_cost}
\title{Compute frozen dinners cost points}
\usage{
frozen_dinners_cost(
  healthier_frozen_dinners_price,
  regular_frozen_dinners_price
)
}
\arguments{
\item{healthier_frozen_dinners_price}{The price of healthier option for frozen dinners. (specifics?)}

\item{regular_frozen_dinners_price}{The price of regular fat frozen dinners.}
}
\value{
The NEMS-S points associated with frozen dinners price.
}
\description{
This function takes in the cost of healthier and regular frozen dinners and compares them to return a NEMS-S score for the cost.
}
\examples{
healthier_frozen_dinners_price <- rnorm(10,4.5,.5)
regular_frozen_dinners_price <- rnorm(10,5.1,.3)
}
