% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ground_beef_avail}
\alias{ground_beef_avail}
\title{Compute the availability points associated with ground beef}
\usage{
ground_beef_avail(lean_beef_varieties)
}
\arguments{
\item{lean_beef_varieties}{The number of lean beef varieties (<10 percent fat).}
}
\value{
The NEMS-S points associated with ground beef availability.
}
\description{
This function takes in the number of varieties of lean and regular fat beef and returns the NEMS-S points associated with availability.
}
\details{
This function implements the scoring method described in Table X of the NEMS-S manual. "Standard ground beef" is 80 percent lean and 20 percent fat by weight. "Lean ground beef" is 90% lean and 10% fat (or less fat percentage) by weight, whichever is available.
}
\examples{
lean_beef_varieties <- sample(1:4, 10)
ground_beef_avail(lean_beef_varieties)
}
