% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{ground_beef_cost}
\alias{ground_beef_cost}
\title{Compute ground beef cost points}
\usage{
ground_beef_cost(lean_beef_price, regular_beef_price)
}
\arguments{
\item{lean_beef_price}{The price of lean ground beef.}

\item{regular_beef_price}{The price of normal fat ground beef. (the specific fat content?)}
}
\value{
the NEMS-S points associated with beef price
}
\description{
This function takes in the cost of lean and regular ground beef and compares them to return a NEMS-S score for the cost.
}
\examples{
lean_beef_price <- rnorm(10,5.8,.5)
regular_beef_price <- rnorm(10,5.5,.3)
ground_beef_cost(lean_beef_price, regular_beef_price)
}
