% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{ground_beef_cost}
\alias{ground_beef_cost}
\title{Compute ground beef cost points}
\usage{
ground_beef_cost(lean_beef_price, regular_beef_price)
}
\arguments{
\item{lean_beef_price}{The price of lean ground beef.}

\item{regular_beef_price}{The price of normal fat ground beef. (the specific fat content?)}
}
\value{
the NEMS-S points associated with beef price
}
\description{
This function takes in the cost of lean and regular ground beef and compares them to return a NEMS-S score for the cost.
}
\details{
This function implements the scoring method described in Table 4.1 of Lundsford (2016). "Standard ground beef" is 80 percent lean and 20 percent fat by weight. "Lean ground beef" is 90 percent lean and 10 percent fat (or less fat percentage) by weight.
}
\examples{
lean_beef_price <- rnorm(10,5.8,.5)
regular_beef_price <- rnorm(10,5.5,.3)
ground_beef_cost(lean_beef_price, regular_beef_price)
}
