% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{juice_drinks_avail}
\alias{juice_drinks_avail}
\title{compute the availability points associated with 100% juice drinks}
\usage{
juice_drinks_avail(healthy_juice_varieties)
}
\arguments{
\item{healthy_juice_varieties}{the number of 100% juice drinks available}
}
\value{
the NEMS-S points associated with availability of healthy juice drinks
}
\description{
compute the availability points associated with 100% juice drinks
}
\examples{
healthy_juice_varieties <- sample(0:3, 10)
juice_drinks_avail(healthy_juice_varieties)
}
