% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{milk_cost}
\alias{milk_cost}
\title{compute milk cost points}
\usage{
milk_cost(lowfat_milk_price, whole_milk_price)
}
\arguments{
\item{lowfat_milk_price}{the price of low fat milk}

\item{whole_milk_price}{the price of whole milk}
}
\value{
the NEMS-S points associated with milk price
}
\description{
compute milk cost points
}
\examples{
lowfat_milk_price <- rnorm(10,2.8,.5)
whole_milk_price <- rnorm(10,3.1,.3)
milk_cost(lowfat_milk_price, whole_milk_price)
}
