% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_scores.R
\name{milk_cost}
\alias{milk_cost}
\title{Compute milk cost points}
\usage{
milk_cost(lowfat_milk_price, whole_milk_price)
}
\arguments{
\item{lowfat_milk_price}{The price of low-fat milk.}

\item{whole_milk_price}{The price of whole milk.}
}
\value{
The NEMS-S points associated with milk price.
}
\description{
This function takes in the cost of low-fat and whole milk and compares them to return a NEMS-S score for the cost.
}
\details{
This function implements the scoring method described in Table 4.1 of Lundsford (2016). "Lowfat milk" is considered skim milk or 1 percent fat, whichever is available.
}
\examples{
lowfat_milk_price <- rnorm(10,2.8,.5)
whole_milk_price <- rnorm(10,3.1,.3)
milk_cost(lowfat_milk_price, whole_milk_price)
}
