% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{soda_avail}
\alias{soda_avail}
\title{compute the availability points associated with diet soda}
\usage{
soda_avail(diet_soda_varieties)
}
\arguments{
\item{diet_soda_varieties}{the number of diet sodas available}
}
\value{
the NEMS-S points associated with availability of diet soda
}
\description{
compute the availability points associated with diet soda
}
\examples{
diet_soda_varieties <- sample(0:3, 10)
soda_avail(diet_soda_varieties)
}
