% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{soda_cost}
\alias{soda_cost}
\title{Compute soda cost points}
\usage{
soda_cost(diet_soda_price, regular_soda_price)
}
\arguments{
\item{diet_soda_price}{The price of regular soda.}

\item{regular_soda_price}{The price of healthier soda.}
}
\value{
The NEMS_S points associated with soda price.
}
\description{
This function takes in the cost of healthier and regular soda and compares them to return a NEMS-S score for the cost.
}
\examples{
diet_soda_price <- rnorm(10,4.1,.5)
regular_soda_price <- rnorm(10,4.1,.3)
}
