% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{soda_cost}
\alias{soda_cost}
\title{compute soda cost points}
\usage{
soda_cost(diet_soda_price, regular_soda_price)
}
\arguments{
\item{diet_soda_price}{the price regular soda}

\item{regular_soda_price}{the price of healthier soda}
}
\value{
the NEMS_S points associated with soda
}
\description{
compute soda cost points
}
\examples{
diet_soda_price <- rnorm(10,4.1,.5)
regular_soda_price <- rnorm(10,4.1,.3)
}
