% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_scores.R
\name{vegetable_avail}
\alias{vegetable_avail}
\title{Compute vegetable availability points}
\usage{
vegetable_avail(varieties_of_vegetables)
}
\arguments{
\item{varieties_of_vegetables}{Number of types of vegetables available.}
}
\value{
The NEMS-S points associated with vegetables availability.
}
\description{
This function takes in the number of varieties of vegetables at a given store and returns the NEMS-S points associated with the availability.
}
\details{
This function implements the scoring method described in Measure 3 of the NEMS-S Protocol.
}
\examples{
varieties_of_vegetables <- sample(1:10, 10)
vegetable_avail(varieties_of_vegetables)


}
