% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nems_s.R
\name{wieners_cost}
\alias{wieners_cost}
\title{Compute wieners cost points}
\usage{
wieners_cost(lean_wieners_price, regular_wieners_price)
}
\arguments{
\item{lean_wieners_price}{The price of hot dogs with less fat (specific fat content?).}

\item{regular_wieners_price}{The price of normal fat hot dogs.}
}
\value{
The NEMS-S points associated with wiener price.
}
\description{
This function takes in the cost of lean and regular fat weiners and compares them to return a NEMS-S score for the cost.
}
\examples{
lean_wieners_price <- rnorm(10,2.5,.3)
regular_wieners_price <- rnorm(10,1.4,.5)
wieners_cost(lean_wieners_price,regular_wieners_price)
}
